+++ frontend-admin/src/pages/LoginPage/LoginPage.styles.ts
@@ -32,43 +32,3 @@ export const loginTitle = css`
   font-size: 3.6rem;
   font-weight: bold;
 `;
-
-export const passwordInput = (theme: Theme) => css`
-  width: 32rem;
-  border: 0.1rem solid ${theme.color.gray};
-  padding: 1.2rem;
-  border-radius: 0.8rem;
-  background-color: white;
-  font-size: 1.6rem;
-  outline: none;
-
-  &:focus {
-    outline: 0.1rem solid black;
-  }
-`;
-
-export const errorMessage = (isError: boolean) => css`
-  visibility: ${isError ? 'visible' : 'hidden'};
-  color: red;
-  font-size: 1.2rem;
-  min-height: 1.2rem;
-  margin-top: 0.8rem;
-  margin: 0.8rem 0 0 1.2rem;
-`;
-
-export const loginButton = (theme: Theme) => css`
-  width: 8rem;
-  padding: 1.2rem;
-  border-radius: 0.8rem;
-  background-color: ${theme.color.peanut400};
-  font-size: 1.6rem;
-  outline: none;
-
-  &:focus {
-    outline: 0.1rem solid black;
-  }
-
-  &:hover {
-    background-color: ${theme.color.peanut500};
-  }
-`;
